Added simplified casting.
Here a cast is really the identity function coupled with a casting of directions.

Casting is for connecting an 'endpoint' coming out of a channel, or a port specified as a paramter in a lambda expresion, to a function application.
We need to allow equivalent types to get connected *and* ensure their directions are compatible.